# PERF: speed up MultiIndex.is_monotonic by 50x